Refactor exceptions of Patient model from Person model
