+++ src/main/java/seedu/address/logic/LogicManager.java
@@ -15,7 +15,7 @@ import seedu.address.logic.parser.AddressBookParser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.person.Person;
+import seedu.address.model.patient.Person;
 import seedu.address.storage.Storage;
 
 /**
