+++ src/main/java/seedu/address/model/ModelManager.java
@@ -11,7 +11,7 @@ import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.model.person.Person;
+import seedu.address.model.patient.Person;
 
 /**
  * Represents the in-memory model of the address book data.
