+++ src/main/java/seedu/address/model/patient/exceptions/PatientNotFoundException.java
@@ -1,0 +1,6 @@
+package seedu.address.model.patient.exceptions;
+
+/**
+ * Signals that the operation is unable to find the specified person.
+ */
+public class PatientNotFoundException extends RuntimeException {}
